Worked on image stitching and viewing
ImageGrabber gets panorama image from streetview and displays it on sphere object for viewing.  PanoShader is a shader to display the image on the inside of the sphere (found at https://medium.com/game-development-stuff/how-to-make-a-360%C2%BA-image-viewer-with-unity3d-b1aa9f99cabb).  TextureScaler is a script that scales textures in unity (petersvp's answer at https://answers.unity.com/questions/150942/texture-scale.html).